Squash linter warnings, move `waitForScrollEnd`
